Fix wall sliding bug.
The bounce factor should apply to both velocity components, not just
in the direction of the bounce. This fixes the wall sliding bug: if
a player has a velocity (10, 0) and they thrust into a wall in the
y-direction, they'll slide along the wall in the x-direction without
losing any speed.

Fixes #31.