Notify user when they've disconnected and provide a button to reconnect.
If the client gets disconnected from the server, pause the simulation
and show a dialog box to that effect. Clicking on the dialog box simply
reloads the page, effectively reconnecting the user.

Without this change, it was hard to tell if you've disconnected. You'd
end up flying around thinking everything was fine because the simulation
was still running.