Split the inheritance hierarchy for *PlayerSprite and *Player.
Before this change, the PlayerSprite classes were inheriting from
their respective Player classes. There were a number of problems
with that approach, not least of which was the idea that a graphical
class was inheriting from a simulation class.

This change clearly separates the two hierarchies out. LocalPlayer
and RemotePlayer both inherit from Player. LocalPlayerSprite and
RemotePlayerSprite both inherit from PlayerSprite. Both Player and
PlayerSprite are root base classes.