Improve simulation accuracy by handling weapons after movement.
Before this change, the sequence of processing was:
 - find out if any weapon was fired
 - update player's position, velocity, and direction
 - send position and weapon data over the network

That sequence resulted in the weapons being fired with the old
player position and velocity but we were sending an updated position
and velocity over the network. So the local client and remote clients
would end up with different simulation states for projectiles.