Fix bug where a remote player "sways" when dotproduct is minimized.
If a remote player minimizes the browser window containing dotproduct,
the game will send position updates at irregular intervals. Sometimes
the interval is > 150 ticks but since we clamp the maximum time delta
to 150 ticks, we end up extrapolating the player's position much less
than it should be and the player sways between the correct location
and the poorly extrapolated location.

Instead of clamping the maximum time delta to 150 ticks, we just
discard position packets that are older than 150 ticks.