Introduce a simulation time. Simulation events are in simulation time.
This change fixes various problems with simulation state going out
of sync between clients when simulation updates are bursty (e.g.
when the browser is minimized). Most notable is the issue of a player
jumping around on the screen because the position timestamps were
generated at the wall clock time of sending the packet instead of at
the simulation time.