Implement repels.
Not exactly an efficient implementation (iterates through all players
and all non-friendly projectiles for every active repel) but a
functioning one nevertheless.

There's a little bit of sketchiness around the repel animation
being displayed out of sync with the simulation. The animation is
shown starting from the time when the repel packet is received
instead of the time the repel was activated.